♻️ Add user relation to deleteTask endpoint

+++ src/tasks/tasks.service.ts
@@ -19,7 +19,7 @@ export class TasksService {
   }
 
   async getTaskbyId(id: number, user: User): Promise<Task> {
-    const task = await this.taskRepository.findOne({ where: { id, userId: user.id} });
+    const task = await this.taskRepository.findOne({ where: { id, userId: user.id } });
 
     if (!task) {
       throw new NotFoundException(`Task with id ${id} not found!`);
@@ -32,8 +32,8 @@ export class TasksService {
     return this.taskRepository.createTask(createTaskDto, user);
   }
 
-  async deleteTaskbyId(id: number): Promise<void> {
-    const result = await this.taskRepository.delete(id);
+  async deleteTaskbyId(id: number, user: User): Promise<void> {
+    const result = await this.taskRepository.delete({ id, userId: user.id });
     if (result.affected === 0) {
       throw new NotFoundException(`Task with id ${id} not found!`);
     }
